noop: API code reorganization
- Moved DataGalaxyApiGlossary to its own file
- Moved utils functions to "datagalaxy_api" file
- Edited imports to avoid breaking changes
- Rename to "list_properties" for shorter name
