+++ tests/test_copy_glossary.py
@@ -1,13 +1,13 @@
-from toolbox.api.datagalaxy_api_workspaces import DataGalaxyApiWorkspace, DataGalaxyApiGlossary, \
-    DataGalaxyBulkResult
+from toolbox.api.datagalaxy_api_glossary import DataGalaxyApiGlossary
+from toolbox.api.datagalaxy_api_workspaces import DataGalaxyApiWorkspace
 from toolbox.commands.copy_glossary import copy_glossary
-from toolbox.api.datagalaxy_api import DataGalaxyApiAuthentication, Token
+from toolbox.api.datagalaxy_api import DataGalaxyApiAuthentication, Token, DataGalaxyBulkResult
 import pytest as pytest
 
 
 # Mocks
 
-def mock_list_glossary_properties_on_source_workspace(self, workspace_name):
+def mock_list_properties_on_source_workspace(self, workspace_name):
     if workspace_name == 'workspace_source':
         return ['property1', 'property1', 'property1']
     return []
@@ -27,10 +27,10 @@ def test_copy_glossary_when_no_property_on_target(mocker):
     workspace_source_mock.return_value = 'workspace_source'
     glossary_properties_on_source_workspace_mock = mocker.patch.object(
         DataGalaxyApiGlossary,
-        'list_glossary_properties',
+        'list_properties',
         autospec=True
     )
-    glossary_properties_on_source_workspace_mock.side_effect = mock_list_glossary_properties_on_source_workspace
+    glossary_properties_on_source_workspace_mock.side_effect = mock_list_properties_on_source_workspace
     bulk_upsert_properties_on_target_workspace_mock = mocker.patch.object(
         DataGalaxyApiGlossary,
         'bulk_upsert_property_tree',
@@ -73,9 +73,9 @@ def test_copy_glossary_same_client_space(mocker):
     workspace_source_mock.return_value = 'workspace_source'
     glossary_properties_on_source_workspace_mock = mocker.patch.object(
         DataGalaxyApiGlossary,
-        'list_glossary_properties',
+        'list_properties',
         autospec=True)
-    glossary_properties_on_source_workspace_mock.side_effect = mock_list_glossary_properties_on_source_workspace
+    glossary_properties_on_source_workspace_mock.side_effect = mock_list_properties_on_source_workspace
     bulk_upsert_properties_on_target_workspace_mock = mocker.patch.object(
         DataGalaxyApiGlossary,
         'bulk_upsert_property_tree',
@@ -121,9 +121,9 @@ def test_copy_glossary_when_workspace_target_does_not_exist(mocker):
     workspace_source_mock = mocker.patch.object(DataGalaxyApiWorkspace, 'get_workspace', autospec=True)
     workspace_source_mock.return_value = None
     glossary_properties_on_source_workspace_mock = mocker.patch.object(DataGalaxyApiGlossary,
-                                                                       'list_glossary_properties',
+                                                                       'list_properties',
                                                                        autospec=True)
-    glossary_properties_on_source_workspace_mock.side_effect = mock_list_glossary_properties_on_source_workspace
+    glossary_properties_on_source_workspace_mock.side_effect = mock_list_properties_on_source_workspace
 
     # ASSERT / VERIFY
     with pytest.raises(Exception, match='workspace workspace_target does not exist'):
