+++ tests/test_main.py
@@ -1,6 +1,5 @@
 import io
 import re
-
 from toolbox.__main__ import run
