+++ toolbox/commands/copy_dataprocessings.py
@@ -1,9 +1,9 @@
 import logging
 from typing import Optional
 
-from toolbox.api.datagalaxy_api import get_access_token, Token
+from toolbox.api.datagalaxy_api import get_access_token, Token, DataGalaxyBulkResult
 from toolbox.api.datagalaxy_api_dataprocessings import DataGalaxyApiDataprocessings
-from toolbox.api.datagalaxy_api_workspaces import DataGalaxyApiWorkspace, DataGalaxyBulkResult
+from toolbox.api.datagalaxy_api_workspaces import DataGalaxyApiWorkspace
 
 
 def copy_dataprocessings(url_source: str,
